requirements.txt
numpy
torch==1.4
torchvision
matplotlib
pillow
terminaltables
tqdm